Refactor field serialization.
`Serializable` class became quite big and I noticed a mistake:
all serialization code is stored inside of the `Serializable` class.
I decided to introduce the interface `IFieldSerializer` to serialize
fields of specific types and moved the logic from `Serializable`
to derived `IFieldSerializer` classes. Now `Serializable` just
goes through all field serializers and uses the one that supports
the field we are currently serializing. Thanks to this we can easily
add support for new field types for serialization so users can make
engine extensions in the future.