Implement referencing external TOML files.
This commit adds functionality to reference external TOML files in
TOML files. The main idea is to reference external node trees or
other serializable objects. Everything is done under the hood so the
end user does not need to care about it. In order for an object
to be referenced in a TOML file the object must be previously
deserialized from the `res` directory. This commit also adds some handy
functions for accesing the `res` directory by using the `ProjectPaths`.

Here is an example: we have a `Node` object and we serialize it
to the `res` directory so that its path will look like this:
`.../res/game/nodeA.toml`. After that, if we deserialize this node,
it will hold the path `game/nodeA.toml` as the path the node was
deserialized from. If we then create a new node B that will reference
this deserialized node A as child node, when we serialize node B,
instead of serializing node A we will store the path `game/nodeA.toml`
and only serialize changed fields of node A (changed compared to the
original/deserialized version of the node). This way we can make
changes to node A and they will appear in the node B. Same goes for
the node trees.

The main idea was to have node trees that reference other node trees,
so that we can have a node tree "Level" and a node tree "Player" and
we can make independent changes to the node tree "Player" and see
the changes in the "Level" node tree.