Make sure all DirectX resources destroyed before the manager.
Checks that we had in Vulkan resource manager for alive resource count
are now also used in DirectX renderer due to the fact that at
destruction DirectX resources usually release their descriptors
but it will not work if the resource manager (that holds descriptor
heaps) is destroyed.

This commit fixes a crash on DirectX renderer destruction.