Delete `ShaderConfiguration` and renderer configuration.
Since they are not used at all. They were supposed to be
used in cases when the renderer has a bunch of macros
that needs to be required on top of the macros that
the pipeline requires. These renderer's macros are
assumed to change during runtime but it turns out we
don't need that right now.